Adds tests for validate_callbacks function
