+++ pyunfold/tests/test_init.py
@@ -4,3 +4,7 @@ import pyunfold
 
 def test_version_exists():
     assert hasattr(pyunfold, '__version__')
+
+
+def test_version_exists():
+    assert hasattr(pyunfold, 'iterative_unfold')
