+++ pyunfold/tests/test_unfold.py
@@ -9,7 +9,6 @@ import pytest
 from .testing_utils import diagonal_response, triangular_response
 
 from pyunfold.unfold import iterative_unfold
-from pyunfold.callbacks import Callback, Logger
 
 PY_VERSION = sys.version_info.major
 
@@ -156,26 +155,3 @@ def test_example_non_square_response():
     unfolded = unfolded[columns]
 
     pd.testing.assert_frame_equal(unfolded, expected)
-
-
-def test_invalid_callbacks_raises():
-    # Run example case
-    data = [100, 150]
-    data_err = [10, 12.2]
-    response = [[0.9, 0.1],
-                [0.1, 0.9]]
-    response_err = [[0.01, 0.01],
-                    [0.01, 0.01]]
-    efficiencies = [0.4, 0.67]
-    efficiencies_err = [0.01, 0.01]
-    with pytest.raises(TypeError) as excinfo:
-        # Specify an invalid callback
-        callbacks = ['not a callback', Logger()]
-        # Perform iterative unfolding
-        iterative_unfold(data, data_err,
-                         response, response_err,
-                         efficiencies, efficiencies_err,
-                         callbacks=callbacks)
-    invalid_callbacks = [c for c in callbacks if not isinstance(c, Callback)]
-    error = ('Found non-callback object in callbacks: {}'.format(invalid_callbacks))
-    assert error == str(excinfo.value)
